% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{as.list.vital}
\alias{as.list.vital}
\title{Set Up a Vital Rates Function With Parameters}
\usage{
\method{as.list}{vital}(x, ...)
}
\arguments{
\item{x}{The function.}

\item{...}{Other variables used by \code{fun}.}
}
\description{
Takes a function and an arbitrary collection of other parameters that are used in that function,
and packages the other parameters into the environment of the function.
}
