% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backmigrate.R
\name{backmeiosis}
\alias{backmeiosis}
\title{Sample from Parental Genotypes}
\usage{
backmeiosis(x, mating, weights, seed, pollen)
}
\arguments{
\item{x}{A vector of genotypes.}

\item{mating}{A mating tensor.}

\item{weights}{A numeric vector of length equal to number of genotypes.}

\item{seed}{A matrix with dimensions \code{length(x)} by number of genotypes.}

\item{pollen}{A matrix with dimensions \code{length(x)} by number of genotypes.}
}
\value{
A \code{length(x)} by 2 matrix whose first column is the genotype index 
and the second column is 1 or 2, meaning:
  1 : seed (first dimension in mating tensor)
  2 : pollen (second dimension in mating tensor)

For genotype=k, the probability 
of choosing parent genotype pair i,j
is proportional to 
  \code{X[i,j] = mating[i,j,k]*seed[i]*pollen[j]}
and given \code{i,j}, the probability of choosing the seed parent is equal to
  \code{weights[i]/(weights[i]+weights[j])} .
}
\description{
This takes a mating tensor, a vector of genotypes, a vector of weights, 
and matrices of numbers of possible seed and pollen parents (with seed[k,j] and pollen[k,j] 
being the numbers of possible j-genotype parents of genotype[k]),
and samples from parental genotypes.
}
