% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backmigrate.R
\name{backmigrate}
\alias{backmigrate}
\title{Sample from Parental Locations}
\usage{
backmigrate(x, migration, weights)
}
\arguments{
\item{x}{A vector of locations (indices).}

\item{migration}{The \code{migration} object.}

\item{weights}{A nonnegative vector of weights.}
}
\value{
A vector of locations of the same form as \code{x}.
If the probability of migrating from i to j is M[i,j], then a sample at j will choose i
with probability equal to \code{weights[i] M[i,j]/sum(weights*M[,j])}.
}
\description{
This takes a set of locations and a nonnegative weighting vector,
and samples from the time-reversed migration probabilities, weighted by the weights.
}
