% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_neighborhoods.R
\name{census_neighborhoods}
\alias{census_neighborhoods}
\title{Construct Census Functions for a Set of Neighborhoods}
\usage{
census_neighborhoods(pop, neighborhoods)
}
\arguments{
\item{pop}{A population object.}

\item{neighborhoods}{A list whose element "neighborhoods" is a SpatialPolygons object.}
}
\value{
A list of functions, one for each genotype.
}
\description{
Given a population object and a set of neighborhoods as returned by \code{sample_neighborhoods},
construct a list of functions, the k-th of which, applied to a matrix of genotype counts form a population object,
will return the vector of total numbers of the k-th genotype occurring in each neighborhood.
}

