% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{crop_population}
\alias{crop_population}
\title{Crop a Population}
\usage{
crop_population(pop, extent, ind.return = FALSE)
}
\arguments{
\item{pop}{A population object.}

\item{extent}{A raster::extent object, or anything that can be used to crop a Raster*.}

\item{ind.return}{Also return the indices of cells in pop$N that are retained?}
}
\value{
A population object, possibly with a new `crop.inds` value.
}
\description{
Crop a population object to a smaller region.
}
