% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{crop_simulation}
\alias{crop_simulation}
\title{Crop a Simulation}
\usage{
crop_simulation(sim, pop, extent)
}
\arguments{
\item{sim}{A simulation object.}

\item{pop}{The corresponding population object.}

\item{extent}{A raster::extent object, or anything that can be used to crop a Raster*.}
}
\value{
A simulation object.  In the new object, `summaries` are NULL.
}
\description{
Crop a simulation object to a smaller region.
}

