% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{demography}
\alias{demography}
\title{Construct a \code{demography} Object.}
\usage{
demography(prob.seed, fecundity, prob.germination, prob.survival,
  pollen.migration, seed.migration, genotypes,
  mating = mating_tensor(genotypes), description = "", ...)
}
\arguments{
\item{prob.seed}{Probability of producing seed, per individual.}

\item{fecundity}{Mean number of seeds produced, per individual, if seeds are produced.}

\item{prob.germination}{Probability of germination (establishment) per seed.}

\item{prob.survival}{Probability of survival per already-extant individual.}

\item{pollen.migration}{A \code{migration} object for pollen dispersal.}

\item{seed.migration}{A \code{migration} object for seed dispersal.}

\item{genotypes}{A character vector of genotypes.}

\item{mating}{An array giving probabilities for offspring genotypes given parental genotypes.}

\item{description}{A description of this object.}

\item{...}{Other parameters that are included verbatim in the output object.}
}
\value{
A \code{demography} S3 object (just a named list).
}
\description{
This puts together the parameters necessary to specify a \code{demography} object,
specifying pollen and seed production and dispersal, and mortality.
The parameters \code{prob.seed}, \code{fecundity}, \code{prob.germination}, 
and \code{prob.survival} can be either Raster* objects, numbers, 
or functions of the current population state (a Raster* object).
}
