% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_kernel.R
\name{discretize_kernel}
\alias{discretize_kernel}
\title{Create a "Discretized" Migration Function from a Continuous One}
\usage{
discretize_kernel(kern, res, radius, sigma = 1, fact = 10)
}
\arguments{
\item{kern}{The migration rate function.}

\item{res}{The resolution of the grid.}

\item{radius}{The maximum distance to extrapolate out to.}

\item{sigma}{Distance scaling for kernel.}

\item{fact}{The number of divisions of each grid cell side when approximating the integral.}
}
\value{
A function as output by `approxfun()`.

The result is designed to be used in migration_matrix, which scales input to the kernel by sigma
and assumes that the kernel is a density function, so multiplies the value by the area of a cell; 
so the resulting function scales as a density function, and takes input in units of sigma.

For good results, the 'fine' grid used in numerical integration should be smaller than sigma;
since the fine grid has a scale of res/fact, you should set fact >= 3 * res / sigma, say.
}
\description{
Given a function kern(x,y) that gives per-individual migration rates and a resolution, 
returns the (approximate) function giving per-individual migration rates *per unit area*
from squares of side length `resolution` centered at (x,y).
}

