% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_distance.R
\name{dispersal_distance}
\alias{dispersal_distance}
\title{Estimate the Root Mean Square Dispersal Distance from a Migration Matrix}
\usage{
dispersal_distance(M, pop, normalized = FALSE)
}
\arguments{
\item{M}{A nonnegative migration matrix, with rows summing to < 1.}

\item{pop}{A population object containing the raster and sets of accessible and habitable cells corresponding to M.}

\item{normalized}{Is the matrix M already row-normalized?}
}
\value{
A vector of mean dispersal distances, for each cell in pop$habitable.
}
\description{
FOr a migration matrix M[i,j], with rowSus(M) all equal to 1,
the root mean square  dispersal distance from location i is 
    sqrt( sum( M[i,] * dist[i,]^2 ) );
this estimates the average dispersal distance across locations
without computing the full distance matrix
by taking a random sample of i and averaging.
}
\details{
If the rows of M sum to something less than 1 (because migrants may be lost),
for the purposes of the migration matrix we should work with
the matrix normalized so that rows do sum to 1,
since we want migrants conditioned on not being lost.
}

