% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{extend_simulation}
\alias{extend_simulation}
\title{Continue a Previously Initiated Simulation}
\usage{
extend_simulation(sim, population, demography, times, append = TRUE,
  N = sim$N[, , dim(sim$N)[3]], tinit = sim$t + 1, summary.times = (sim$t
  + 1):max(times), summaries = NULL, ...)
}
\arguments{
\item{sim}{The previous \code{simulation} object.}

\item{population}{A \code{population} object, the initial state (t=0) of the population.}

\item{demography}{A \code{demography} object, containing the below parameters.}

\item{times}{A sorted integer vector of times to record the full state of the population at.}

\item{append}{If TRUE, append results to results in previous simulation object.}

\item{N}{Overrides the matrix of genotype counts in \code{population}, if present.}

\item{tinit}{The current time at the beginning of the simulation (defaults to t=0).}

\item{summary.times}{A sorted integer vector of times to record summaries of the population at.}

\item{summaries}{A list of functions to apply to the matrix of genotype counts, population$N, at generations \code{summary.times}.}

\item{...}{Additional parameters that will be passed to \code{generation()}.}
}
\value{
A simulation object, as \code{simulation}.
If summary functions are not the same, the initial ones are removed.
}
\description{
Run a previously initiated popluation for more generations.
}

