% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generation}
\alias{generation}
\title{Simulate a Single Generation.}
\usage{
generation(population, demography, N = population$N,
  prob.seed = demography$prob.seed, fecundity = demography$fecundity,
  prob.germination = demography$prob.germination,
  prob.survival = demography$prob.survival,
  pollen.migration = demography$pollen.migration,
  seed.migration = demography$seed.migration,
  genotypes = demography$genotypes, mating = demography$mating,
  death.first = (is.null(demography$death.first) || demography$death.first),
  expected = FALSE, return.everything = FALSE, debug = FALSE, ...)
}
\arguments{
\item{population}{A \code{population} object, the initial state of the population.}

\item{demography}{A \code{demography} object, containing the below parameters.}

\item{N}{Optionally, a matrix of the same form as \code{population$N} (overrides \code{population$N} if present).}

\item{prob.seed}{Probability of producing seed, per individual.}

\item{fecundity}{Mean number of seeds produced, per individual, if seeds are produced.}

\item{prob.germination}{Probability of germination (establishment) per seed.}

\item{prob.survival}{Probability of survival per already-extant individual.}

\item{pollen.migration}{A \code{migration} object for pollen dispersal.}

\item{seed.migration}{A \code{migration} object for seed dispersal.}

\item{genotypes}{A character vector of genotypes.}

\item{mating}{An array giving probabilities for offspring genotypes given parental genotypes.}

\item{death.first}{If TRUE, germination probability is a function of densities *after* death has taken place.}

\item{expected}{If TRUE, do no sampling, returning only expected values.}

\item{return.everything}{If TRUE, returns the number of seeders, amount of pollen, density of seeds, number of germinating seeds, and number of dying individuals.}

\item{debug}{Do checks for NAs and the like.}

\item{...}{Additional parameters that will be passed to demographic functions.}
}
\value{
A matrix of the same dimensions as \code{population$N},
unless \code{return.everything} is TRUE, in which case it is a list with entries:

  seeders : number of seeding individuals
  pollen : quantity of pollen produced, by genotype of parent
  seed.production : quantity of fertilized seeds produced
  seeds.dispersed : quantity of seeds, post-dispersal
  germination : number of new seeds that germinate (survive)
  death : number of previously alive individuals that die
}
\description{
This steps a population forwards by one generation,
according to the parameters in a \code{demography} object,
(which can also be passed in separately; see \code{demography}).
}
