% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generation_raster}
\alias{generation_raster}
\title{Apply a \code{demography} Object to a Raster}
\usage{
generation_raster(population, demography, prob.seed = demography$prob.seed,
  fecundity = demography$fecundity,
  prob.germination = demography$prob.germination,
  prob.survival = demography$prob.survival,
  pollen.migration = demography$pollen.migration,
  seed.migration = demography$seed.migration,
  genotypes = demography$genotypes, mating = demography$mating, ...)
}
\arguments{
\item{population}{A Raster* object.}

\item{demography}{A \code{demography} object, containing the below parameters.}

\item{prob.seed}{Probability of producing seed, per individual.}

\item{fecundity}{Mean number of seeds produced, per individual, if seeds are produced.}

\item{prob.germination}{Probability of germination (establishment) per seed.}

\item{prob.survival}{Probability of survival per already-extant individual.}

\item{pollen.migration}{A \code{migration} object for pollen dispersal.}

\item{seed.migration}{A \code{migration} object for seed dispersal.}

\item{genotypes}{A character vector of genotypes.}

\item{mating}{An array giving probabilities for offspring genotypes given parental genotypes.}

\item{...}{Additional parameters that will be passed to demographic functions.}
}
\value{
A new population of the same form as the old.
}
\description{
This steps a population, stored as a Raster*, forwards by one generation,
according to the parameters in a \code{demography} object,
(which can also be passed in separately; see \code{demography}).
}

