% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_matrix.R
\name{geo_power}
\alias{geo_power}
\title{Take a Geometric Power of a Matrix}
\usage{
geo_power(M, p, eps = 1e-08, n = log(eps)/log(p))
}
\arguments{
\item{M}{The migration matrix.}

\item{p}{The parameter in the geometric distribution (must be between 0 and 1).}

\item{eps}{The numerical tolerance.}

\item{n}{The number of terms to use, minus one (defaults so that p^n = eps).}
}
\value{
A matrix of the same form as M.
}
\description{
For a matrix M, compute (1-p) ( I + p M + p^2 M^2 + ... ) = (1-p)(I-pM)^{-1} .
If M is a transition matrix, the result is the distribution after taking a Geometric(p) number of steps.
}

