% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_N}
\alias{get_N}
\title{Access Values of Habitable Cells in a Population}
\usage{
get_N(x, i, j)
}
\arguments{
\item{x}{The population object.}

\item{i}{The rows to read values of, indexed by cells in x$habitat.}

\item{j}{The columns to read values of, or names of genotypes.}
}
\description{
Provides a method to read values to the matrix of population sizes
in a population object
indexed by cells in the underlying Raster* rather than rows the matrix itself,
which records only values at habitable locations.
}

