% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{getElement.vital}
\alias{$.vital}
\alias{$<-.vital}
\alias{[.vital}
\alias{[[.vital}
\alias{[[<-.vital}
\alias{getElement.vital}
\title{Assign or Extract Items from the Environment of a Vital Function.}
\usage{
getElement.vital(object, name)

\method{$}{vital}(x, name)

\method{$}{vital}(x, i) <- value

\method{[[}{vital}(x, i) <- value

\method{[[}{vital}(x, i)

\method{[}{vital}(x, i)
}
\arguments{
\item{x, object}{The vital function.}

\item{i, name}{The name of the item to extract.}

\item{value}{The new value to assign.}
}
\description{
Extraction methods return NULL if the item is not defined, just like for lists.
}
\section{Methods (by generic)}{
\itemize{
\item \code{$}: Extract items from the environment of a vital function.

\item \code{$<-}: Assign to environment of a vital function.

\item \code{[[<-}: Assign to environment of a vital function.

\item \code{[[}: Extract items from the environment of a vital function.

\item \code{[}: Extract multiple items from the environment of a vital function.
}}

