% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_adjacency.R
\name{layer_adjacency}
\alias{layer_adjacency}
\title{Adjacency Matrix from a Raster Layer.}
\usage{
layer_adjacency(layer, from = which(raster::values(!is.na(layer))),
  to = from, directions = 4, normalize = FALSE)
}
\arguments{
\item{layer}{A Raster* object.}

\item{from}{The indices of the cells corresponding to rows in the output matrix. [default: non-NA cells]}

\item{to}{The indices of the cells corresponding to columns in the output matrix. [default: same as \code{from}]}

\item{directions}{the number of directions in which cells should be
connected: 4 (rook's case), 8 (queen's case), 16 (knight and
one-cell queen moves), or 'bishop' to connect cells with
one-cell diagonal moves. Or a neigborhood matrix (see \code{raster::adjacent}).
[default: 4]}

\item{normalize}{If \code{TRUE}, normalize rows to sum to 1.}
}
\value{
A sparse Matrix \code{M}, where \code{M[i,j]} is equal to 1 if cell \code{from[i]} is adjacent to \code{to[j]}, and is 0 otherwise.
}
\description{
This returns the (sparse) Matrix giving the adjacency matrix of cells in \code{layer}.
}
\keyword{layers}
