% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_generation.R
\name{lineage_generation}
\alias{lineage_generation}
\title{Simulate a Single Backwards Generation}
\usage{
lineage_generation(lineages, N, gen, num.alleles, demog, debug = FALSE)
}
\arguments{
\item{lineages}{A two-column matrix with columns labeled 'location' and 'genotype'.}

\item{N}{The population state at the start of the previous generation.}

\item{gen}{A list as returned by \code{generation(...,return.everything=TRUE)}.}

\item{num.alleles}{A vector giving the number of alleles of the type followed in each genotype.}

\item{demog}{The demography object that contains migration matrices used to make the simulation.}

\item{debug}{Whether to run sanity checks.}
}
\value{
A matrix of the same form as \code{lineages}.
}
\description{
Given a collection of lineages, all of the same genotype,
sample from their location in the previous generation,
given the entire set of reproductive numbers in that generation.
}

