% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_neighborhoods.R
\name{make_circles}
\alias{make_circles}
\title{Make Circles Centered at a Collection of Coordinates}
\usage{
make_circles(centers, radii, proj4string = sp::CRS(sp::proj4string(centers)))
}
\arguments{
\item{centers}{The centers of the circles.}

\item{radii}{The radii of the circles.}

\item{proj4string}{The proj4string.}
}
\value{
A SpatialPolygons object of circles.
}
\description{
Make Circles Centered at a Collection of Coordinates
}
