% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{make_population}
\alias{make_population}
\title{Set Up a Population from a Configuration}
\usage{
make_population(habitat, inaccessible.value, uninhabitable.value, genotypes,
  N = 0, extent)
}
\arguments{
\item{habitat}{A RasterLayer, or the path to the file where the raster is stored.}

\item{inaccessible.value}{The values in the raster that should be marked as inaccessible.}

\item{uninhabitable.value}{The values in the raster that should be marked as not habitable.}

\item{genotypes}{A character vector of genotypes.}

\item{N}{Numerical value(s) to initialize the matrix of genotype counts with.}

\item{extent}{A raster::extent object (or a vector of xmin,xmax,ymin,ymax) to crop the habitat with.}
}
\description{
Provides a method to read and store population configuration.
}
