% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mating_tensor.R
\name{mating_tensor}
\alias{mating_tensor}
\title{Comptue the Mating Tensor for a Set of Genotypes.}
\usage{
mating_tensor(maternal, paternal = maternal, offspring = maternal,
  sep = "", phased = FALSE)
}
\arguments{
\item{maternal}{Character vector of genotypes.}

\item{paternal}{Character vector of genotypes. [default: same as maternal]}

\item{offspring}{Character vector of genotypes. [default: same as maternal]}

\item{sep}{Character separating the two alleles. [default: '']}

\item{phased}{Does order matter for the diploid genotypes?}
}
\value{
A three-dimensional array, with dimnames.
If \code{phased} is \code{FALSE} then the two paternal alleles are sorted before being combined together into the offspring diploid genotype.
}
\description{
This returns the three-dimensional array whose \code{[i,j,k]}th entry is the probability that maternal genotype \code{i} and paternal genotype \code{j} combine to make offspring genotype \code{k}
}

