% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Migrate the Quantities in a Population.}
\usage{
migrate(x, migration)
}
\arguments{
\item{x}{A numeric matrix, or a population object, in which case migration will be applied to \code{x$N}.}

\item{migration}{The \code{migration} object.}
}
\value{
A matrix of values  of the same form as \code{x}.  See \code{migrate} for details.
}
\description{
This applies a \code{migration} operation to a given population object.
}
