% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{migration}
\alias{migration}
\title{Construct a \code{migration} Object.}
\usage{
migration(kern, radius, sigma = 1, normalize = 1, n.weights = 1,
  do.M = FALSE, population, ...)
}
\arguments{
\item{kern}{A function of scaled distance giving relative weights. (or, "gaussian", "cauchy")}

\item{radius}{The maximum distance to migrate over.}

\item{sigma}{A number that distances are scaled by before being passed to \code{kern}.}

\item{normalize}{Total number of migrants per unit of input.}

\item{n.weights}{Weights on the number of times to apply the smoother: the resulting operator is x -> (1-sum(n.weights)) * x + n.weights[1] * M x + n.weights[2] * M^2 x + ...}

\item{do.M}{Precompute an explicit migration matrix?}

\item{population}{A \code{population}; used in computing \code{M}.}

\item{...}{Other parameters that are included verbatim in the output object.}

\item{from}{Used in computing \code{M}.}

\item{to}{Used in computing \code{M}.}
}
\value{
A \code{migration} S3 object (just a named list).

If \code{kern} is a migration object, can be used to add a migration matrix to it.
See \code{migration_matrix}.

Since the memory usage of the migration matrix increases quadratically with the size of the radius,
it can be useful to keep it small and instead apply the migration operator several times;
this is achieved by setting \code{n.weights} to have weight at indices larger than 2.
}
\description{
This puts together the parameters necessary to specify a \code{migration} object,
specifying the form of the dispersal and the total number of dispersers.
}

