% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{names.vital}
\alias{names.vital}
\title{Extract Names of Objects in Environment of a vital Function}
\usage{
\method{names}{vital}(x)
}
\arguments{
\item{x}{The vital function.}
}
\description{
Extract Names of Objects in Environment of a vital Function
}
