% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{nhabitable}
\alias{nhabitable}
\title{Number of Habitable Cells in a Population.}
\usage{
nhabitable(population)
}
\arguments{
\item{population}{The population object.}
}
\description{
Returns the number of habitable cells in a population.
}
