% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.simulation}
\alias{plot.simulation}
\title{Animate a Simulation}
\usage{
plot.simulation(sim, pop, max.frames = 300, zlim = range(sim$N, finite =
  TRUE), pause = interactive(), legend.width = 2, legend.mar = 12,
  animate = NULL, cleanup = FALSE, duration = 10, height = 2,
  width = length(pop$genotypes) * 2, res = 300, pointsize = 10, ...)
}
\arguments{
\item{sim}{A simulation object.}

\item{pop}{The corresponding population object.}

\item{max.frames}{Maximum number of frames to plot.}

\item{zlim}{Range of values to represent by colors, consistently.}

\item{pause}{Pause for user input between frames? (left click to continue, right click to stop)}

\item{legend.width, legend.mar}{Parameters controlling how the legend is displayed.}

\item{animate}{If not FALSE, instead of plotting the result, package frames into an mp4 animation with this file name.}

\item{cleanup}{Whether to delete pngs produced if 'animate' is TRUE.}

\item{duration}{Duration of the animation in seconds.}

\item{...}{Additional parameters passed to plot( ).}
}
\description{
Sequentially plot the time steps in a simulation, optionally cropped.
}

