% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pop_to_raster}
\alias{pop_to_raster}
\title{Convert a population to a RasterBrick.}
\usage{
pop_to_raster(population, x = population$N)
}
\arguments{
\item{population}{The population object.}

\item{x}{The values to be put into the Raster.}
}
\description{
Puts a set of values into the map given by the habitat of the population.
}
