% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Construct a \code{population} Object.}
\usage{
population(habitat, accessible = !is.na(raster::values(habitat)),
  habitable = !is.na(raster::values(habitat)), genotypes = colnames(N),
  N = matrix(0, nrow = sum(habitable), ncol = length(genotypes)),
  description = "", ...)
}
\arguments{
\item{habitat}{RasterLayer with values that can be used in other computations.}

\item{accessible}{Logical vector indicating which cells in `habitat` that migrants will attempt to go to.}

\item{habitable}{Locgical vector indicating which cells in `habitat` that may have positive population.}

\item{genotypes}{Character vector of the genotypes.}

\item{N}{Matrix indexed by (habitable cells) x (genotypes) giving the number of each genotype in each habitable cell.}

\item{description}{A description of this object.}

\item{...}{Other parameters that are included verbatim in the output object.}
}
\value{
A \code{population} S3 object (just a named list).
}
\description{
This puts together the parameters necessary to specify a \code{population} object,
specifying the reference map (as a Raster), accessible and habitable locations,
the genotypes, and the census at each habitable location.
}

