% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_habitat.R
\name{random_habitat}
\alias{random_habitat}
\title{Simulate a Random Environment}
\usage{
random_habitat(diam = 20000, res = 100, randfun = function(n) pmin(20, (2
  + rcauchy(n))), kern = "gaussian", sigma = 300, radius = 1500)
}
\arguments{
\item{diam}{Diameter of the desired habitat (in real units, not number of cells).}

\item{res}{Spatial resolution of the Raster.}

\item{randfun}{Function used to sample values for the habitat (negative values will be set to NA).}

\item{kern}{Smoothing kernel.}

\item{sigma}{Smoothing kernel scale.}

\item{radius}{Smoothing kernel maximum radius.}
}
\value{
A RasterLayer with nonnegative and missing values.
}
\description{
Simulates a random environment, with holes and varying carrying capacity.
}

