% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinom_raster.R
\name{rbinom_matrix}
\alias{rbinom_matrix}
\title{Sample from a Binomial, Preserving Dimension}
\usage{
rbinom_matrix(size, prob)
}
\arguments{
\item{size}{Integer vector, matrix, or array.}

\item{prob}{Numeric vector, matrix, or array.}
}
\value{
An integer object of the same dimensions as the input (recycled to match each other),
and the same dimnames if these can be matched.
}
\description{
Works as `rbinom`, except that it preserves the dimensions, and dimnames of the input(s).
}

