% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinom_raster.R
\name{rbinom_raster}
\alias{rbinom_raster}
\title{Sample Binomial Values in a Raster* or Population}
\usage{
rbinom_raster(size, prob, only.values = FALSE)
}
\arguments{
\item{size}{Number, or Raster* object of sample sizes (must be integers).}

\item{prob}{Number, or Raster* object of probabilities.}

\item{only.values}{Return only the vector of non-NA values?}
}
\value{
A Raster* object of the same form as the input,
or if \code{only.values} is \code{TRUE} the vector of values.
At least one of \code{size} or \code{prob} must be Raster* objects.
}
\description{
Returns a Raster* object whose values are random, Binomial(N,prob.seed),
or just the corresponding values.
}
