% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpois_raster.R
\name{rpois_matrix}
\alias{rpois_matrix}
\title{Sample from a Poisson, Preserving Dimension}
\usage{
rpois_matrix(lambda)
}
\arguments{
\item{lambda}{A numeric object whose values can be assigned to as "lambda[] <- ...".}
}
\value{
An object of the same form as the input, with integer values.
}
\description{
Works as \code{rpois}, but returns an object of the same form as the input.
}
