% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpois_raster.R
\name{rpois_raster}
\alias{rpois_raster}
\title{Sample Poisson Raster}
\usage{
rpois_raster(lambda, only.values = FALSE)
}
\arguments{
\item{lambda}{Raster* object of means.}

\item{only.values}{Return only the vector of non-NA values?}
}
\value{
A Raster* object of the same form as the input,
or if \code{only.values} is \code{TRUE} the vector of values.
}
\description{
Returns a Raster* which is Poisson-values with mean given by the input Raster*.
}

