% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_neighborhoods.R
\name{sample_neighborhoods}
\alias{sample_neighborhoods}
\title{Sample Neighborhoods in a Raster*}
\usage{
sample_neighborhoods(x, n, radius, separation = radius, thin = TRUE)
}
\arguments{
\item{x}{The Raster* object.}

\item{n}{The number of points to sample.}

\item{radius}{The radius of the neighborhoods.}

\item{separation}{The minimum distance separating neighborhood centers (if \code{thin} is TRUE).}

\item{thin}{Whether to thin neighborhoods.}
}
\value{
A named list containing:
   centers = a SpatialPoints object of centers
   neighborhoods = SpatialPolygons object of the circle(s)
   center.cells = the indices of the cells of the Ratser* object that contain the centers
}
\description{
Chooses a set of randomly sampled points centered on non-NA cells of the input RasterLayer,
and returns a randomly chosen maximal subset of nonoverlapping circles centered at these points.
}

