% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_production.R
\name{seed_production_raster}
\alias{seed_production_raster}
\title{Mean Seed Production}
\usage{
seed_production_raster(seeders, pollen, mating, fecundity = 1)
}
\arguments{
\item{seeders}{A Raster* of numbers of seeding individuals.}

\item{pollen}{A Raster* of pollen density.}

\item{mating}{An array with probabilities of producing each genotype from each parental genotype.}

\item{fecundity}{Scaling factor that multiplies the resulting matrix.}
}
\value{
A Raster* of the same form as the input.
}
\description{
Find the Raster* of mean seed production, by genotype,
given the local numbers of seeding individuals and pollen density,
by genotype.
}

