% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_clumps.R
\name{select_clumps}
\alias{select_clumps}
\title{Divide a Raster into Regions Separated by a Minimum Distance}
\usage{
select_clumps(x, threshold, ...)
}
\arguments{
\item{x}{The Raster* object.}

\item{threshold}{The minimum distance defining the groups.}

\item{...}{Additional parameters passed to \code{raster::clump}.}
}
\value{
A Raster* of the same form as \code{x}.
}
\description{
Classifies the non-NA values in a Raster* object into the unqiue minimal sets
such that if two cells are within a given threshold distance of each other,
they are in the same set.
}

