% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{setup_demography}
\alias{setup_demography}
\title{Set up a Demography Object for a Concrete Population}
\usage{
setup_demography(demog, pop, ...)
}
\arguments{
\item{demog}{The demography object.}

\item{pop}{The population object.}

\item{...}{Other parameters passed to \code{migration()}.}
}
\value{
A \code{demography} object, as before, but with but with any migration objects having a migration matrix.
}
\description{
Demography objects are agnostic about the map they work on,
but for application we need to link them to a particular one.
This function does that, by setting up any elements that are \code{migration} or \code{vital} objects.
}

