% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{setup_migration}
\alias{setup_migration}
\title{Set up Migration Object for a Concrete Population}
\usage{
setup_migration(m, population, ...)
}
\arguments{
\item{m}{The migration object.}

\item{population}{The population object.}

\item{...}{Other parameters passed to \code{migration()}.}
}
\value{
A \code{migration} object, as before, but with but with a migration matrix (see \code{migration()}).
}
\description{
Migration objects are agnostic about the map they work on,
but for application we need to add a migration matrix.
}
