% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{setup_vital}
\alias{setup_vital}
\title{Set up a Vital Function for a Concrete Population}
\usage{
setup_vital(fun, pop, ...)
}
\arguments{
\item{fun}{The vital function.}

\item{pop}{The population object.}

\item{...}{Other parameters passed to \code{migration()}.}
}
\value{
A \code{vital} function, as before, but with any migration objects having a migration matrix.
}
\description{
A vital function may include migration objects,
which for use we need to link to particular population
(precomputing the migration matrix).
*In addition*, every element of the population is added
to the environment of the function,
so that the function can depend on objects only visible in the population.
}
