% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_pop}
\alias{simulate_pop}
\alias{simulation}
\title{Simulate Many Generations}
\usage{
simulate_pop(population, demography, times, N = population$N, tinit = 0,
  summary.times = 1:max(times), summaries = NULL, stop.fun = NULL, ...)

simulation(N, summaries, times, summary.times, t, ...)
}
\arguments{
\item{population}{A \code{population} object, the initial state (t=0) of the population.}

\item{demography}{A \code{demography} object, containing the below parameters.}

\item{times}{A sorted integer vector of times to record the full state of the population at.}

\item{N}{Overrides the matrix of genotype counts in \code{population}, if present.}

\item{tinit}{The current time at the beginning of the simulation (defaults to t=0).}

\item{summary.times}{A sorted integer vector of times to record summaries of the population at.}

\item{summaries}{A list of functions to apply to the matrix of genotype counts, population$N, at generations \code{summary.times}.}

\item{stop.fun}{A function applied to N that, if it returns TRUE, will stop the simulation. (remainder of output array will be entirely zeros)}

\item{...}{Additional parameters that will be passed to \code{generation()}.}
}
\value{
A named list, with elements
      N = an array with dimensions as \code{population$N} x \code{length(times)}.
      summaries = A list of matrices, with dimensions (length of summary.times) x (length of output of summary function).
      times = As in the input.
      summary.times = As in the input.
      t = The final generation.
}
\description{
This steps a population forwards by many generations,
according to the parameters in a \code{demography} object,
returning the result at specified times in an array.
}
\section{Functions}{
\itemize{
\item \code{simulation}: Constructor for a simulation object.
}}

