% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_matrix.R
\name{subset_migration}
\alias{subset_migration}
\title{Subset a Migration Matrix to Match a Smaller Raster of the Same Resolution}
\usage{
subset_migration(M, old, new, from.old = which(!is.na(raster::values(old))),
  to.old = from.old, from.new = which(!is.na(raster::values(new))),
  to.new = from.new)
}
\arguments{
\item{M}{The migration matrix.}

\item{old}{The original Raster* the migration matrix was computed for.}

\item{new}{The Raster* for the new migration matrix.}

\item{from.old}{The indices of the "from" cells corresponding to rows in the original migration matrix.}

\item{to.old}{The indices of the "to" cells corresponding to columns in the original migration matrix.}

\item{from.new}{The indices of the "from" cells corresponding to rows in the resulting migration matrix.}

\item{to.new}{The indices of the "to" cells corresponding to columns in the resulting migration matrix.}
}
\value{
A migration matrix.  See \code{migrate}.
}
\description{
This converts a migration matrix computed for one raster
into a migration matrix for another raster that must be a subset of the first.
This does not do any renormalization.
}

