% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{values.population}
\alias{values.population}
\title{Extract Matrix of Genotype Counts}
\usage{
values.population(pop)
}
\arguments{
\item{pop}{The population object.}
}
\value{
The matrix of genotype counts.
}
\description{
Parallels the \code{values()} function of the raster package,
which allows usage of the same code for both types in some situations.
}

