% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{vital}
\alias{vital}
\title{Set Up a Vital Rates Function With Parameters}
\usage{
vital(fun, ...)
}
\arguments{
\item{fun}{The function, whose first argument is \code{N}, a matrix of genotype counts.}

\item{...}{Other variables used by \code{fun}.}
}
\description{
Takes a function and an arbitrary collection of other parameters that are used in that function,
and packages the other parameters into the environment of the function.
}
