% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_matrix.R
\name{write_triplets}
\alias{write_triplets}
\title{Write out a Sparse Matrix in Triplet Form}
\usage{
write_triplets(M, file, row.names = FALSE, ...)
}
\arguments{
\item{M}{The matrix.}

\item{file}{The file name to write out to.}

\item{row.names}{Whether to write out row indices.}

\item{...}{Additional parameters to be passed to \code{write.table}.}
}
\value{
The result of \code{write.table}.
}
\description{
Writes out a file (using \code{write.table}) containing columns
 row  col   value
  i    j     M[i,j]
for every nonzero entry of M.
}

