% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_habitat.R
\name{flat_habitat}
\alias{flat_habitat}
\title{Create a Flat Habitat}
\usage{
flat_habitat(diam = 20000, res = 100, value = 1, width = diam,
  height = diam)
}
\arguments{
\item{diam}{Diameter of the desired habitat (in real units, not number of cells).}

\item{res}{Spatial resolution of the Raster.}

\item{value}{The value(s) to populate the habitat with (will be recycled).}

\item{width}{Width of the region (default: diam).}

\item{height}{Height of the region (default: diam).}
}
\value{
A RasterLayer with nonnegative and missing values.
}
\description{
Simulates a random environment, with holes and varying carrying capacity.
}
