% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{plot.population}
\alias{plot.population}
\title{Plot a Population}
\usage{
plot.population(pop, zlim = range(pop$N, finite = TRUE), type, ...)
}
\arguments{
\item{pop}{The population object.}

\item{zlim}{The range of color values.}

\item{type}{Either "raster" (default for 2d), or "line" (default for 1d).}

\item{...}{Other parameters passed to plot().}
}
\description{
Plots a population object using pop_to_raster() and raster's method for plotting RasterBrick objects.
}

