% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_simulation.R
\name{simulate_lineages}
\alias{simulate_lineages}
\title{Simulate Lineages}
\usage{
simulate_lineages(lineages, gens, num.alleles, demog, ...)
}
\arguments{
\item{lineages}{A two-column matrix with columns labeled 'location' and 'genotype'.}

\item{gens}{A list as returned by \code{simulate_pop(...,return=everything=TRUE)}.}

\item{num.alleles}{A vector giving the number of alleles of the type followed in each genotype.}

\item{demog}{The demography object that contains migration matrices used to make the simulation.}

\item{...}{Other parameters passed to \code{lineage_generation()}.}
}
\value{
An array of dimensions \code{c(dim(lineages),length(gens$times))}, 
in the SAME order as gens$times, so that if x is the output,
  x[,,1] records lineage locations at the *most distant* time,
  x[,1,1] gives the index of the spatial locations, and
  x[,2,1] gives the index of the genotypes of the lineages.
See plot_lineages for how to recover coordinates of the lineages.
}
\description{
Given a collection of lineages, all of the same genotype,
sample from their locations across previous generations,
given the entire set of reproductive numbers in those generation.
}
